transitioned applicationinsights-js
