+++ applicationinsights-js/applicationinsights-js-tests.ts
@@ -1,5 +1,6 @@
-///<reference path='applicationinsights-js.d.ts' />
-// More samples on: https://github.com/Microsoft/ApplicationInsights-JS/blob/master/API-reference.md 
+// More samples on: https://github.com/Microsoft/ApplicationInsights-JS/blob/master/API-reference.md
+
+/// <reference types="applicationinsights-js" />
 
 var config: Microsoft.ApplicationInsights.IConfig = {
     instrumentationKey: "<your iKey>",
